requirements.txt
numpy
Pillow
keras==2.8.0
tensorflow==2.8.0
flask
Werkzeug
gunicorn